### 2022-06-09 #4462 @editorialbot
[PRE REVIEW]: MSG: A software package for interpolating stellar spectra in pre-calculated grids
**Submitting author:** @rhdtownsend (<a href="http://orcid.org/0000-0002-2522-8605">Richard Townsend</a>)
**Repository:** https://github.com/rhdtownsend/msg
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0
