### 2022-06-09 #4462 @kyleniemeyer
Hi @rhdtownsend, we will use this pre-review issue to find an editor and reviewers for your submission. Any recommendations you might have for the latter would be welcome.

As we are getting started, can you take a look at those missing DOIs from your references?
