+++ paper.bib
@@ -5,7 +5,8 @@
 	pages = {A25},
 	title = {{A collection of model stellar spectra for spectral types B to early-M}},
 	volume = {618},
-	year = 2018}
+	year = 2018,
+	doi = {10.1051/0004-6361/201732484}}
 
 @inproceedings{Allende-Prieto:2015,
        author = {{Allende-Prieto}, Carlos and {Apogee Team}},
@@ -23,7 +24,8 @@
 	pages = {A111},
 	title = {{Determining stellar atmospheric parameters and chemical abundances of FGK stars with iSpec}},
 	volume = {569},
-	year = 2014}
+	year = 2014,
+	doi = {10.1051/0004-6361/201423945}}
 
 @article{Chiavassa:2018,
 	author = {{Chiavassa}, A. and {Casagrande}, L. and {Collet}, R. and {Magic}, Z. and {Bigot}, L. and {Th{\'e}venin}, F. and {Asplund}, M.},
@@ -32,7 +34,8 @@
 	pages = {A11},
 	title = {{The STAGGER-grid: A grid of 3D stellar atmosphere models. V. Synthetic stellar spectra and broad-band photometry}},
 	volume = {611},
-	year = 2018}
+	year = 2018,
+	doi = {10.1051/0004-6361/201732147}}
 
 @article{de-Laverny:2012,
 	author = {{de Laverny}, P. and {Recio-Blanco}, A. and {Worley}, C.~C. and {Plez}, B.},
@@ -41,7 +44,8 @@
 	pages = {A126},
 	title = {{The AMBRE project: A new synthetic grid of high-resolution FGKM stellar spectra}},
 	volume = {544},
-	year = 2012}
+	year = 2012,
+	doi = {10.1051/0004-6361/201219330}}
 
 @article{Husser:2013,
 	author = {{Husser}, T. -O. and {Wende-von Berg}, S. and {Dreizler}, S. and {Homeier}, D. and {Reiners}, A. and {Barman}, T. and {Hauschildt}, P.~H.},
@@ -50,7 +54,8 @@
 	pages = {A6},
 	title = {{A new extensive library of PHOENIX stellar atmospheres and synthetic spectra}},
 	volume = {553},
-	year = 2013}
+	year = 2013,
+	doi = {10.1051/0004-6361/201219058}}
 
 @article{Kirby:2011,
 	author = {{Kirby}, Evan N.},
@@ -60,7 +65,8 @@
 	pages = {531},
 	title = {{Grids of ATLAS9 Model Atmospheres and MOOG Synthetic Spectra}},
 	volume = {123},
-	year = 2011}
+	year = 2011,
+	doi = {10.1086/660019}}
 
 @article{Lanz:2003,
 	author = {{Lanz}, T. and {Hubeny}, I.},
@@ -69,7 +75,8 @@
 	pages = {417-441},
 	title = {{A Grid of Non-LTE Line-blanketed Model Atmospheres of O-Type Stars}},
 	volume = 146,
-	year = 2003}
+	year = 2003,
+	doi = {10.1086/374373}}
 
 @article{Lanz:2007,
 	author = {{Lanz}, T. and {Hubeny}, I.},
@@ -78,7 +85,8 @@
 	pages = {83-104},
 	title = {{A Grid of NLTE Line-blanketed Model Atmospheres of Early B-Type Stars}},
 	volume = 169,
-	year = 2007}
+	year = 2007,
+	doi = {10.1086/511270}}
 
 @article{Meszaros:2013,
 	author = {{M{\'e}sz{\'a}ros}, Sz. and {Allende Prieto}, C.},
@@ -88,7 +96,8 @@
 	pages = {3285-3291},
[…]
