### 2022-06-27 #4462 @rhdtownsend
Sure —

Carlos Allende Prieto — https://github.com/callendeprieto
Dan Forman-Mackey — https://github.com/dfm (also a JOSS editor, I believe)
Hugo Tabernero — https://github.com/hmtabernero
Ansgar Wehrhahn — https://github.com/AWehrhahn

cheers,

Rich
