### 2022-06-27 #4462 @arfon
@callendeprieto @hmtabernero  @AWehrhahn – would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The software under consideration is

[_MSG: A software package for interpolating stellar spectra in pre-calculated grids_](https://github.com/rhdtownsend/msg)
