### 2022-07-21 #4602 @editorialbot
[REVIEW]: MSG: A software package for interpolating stellar spectra in pre-calculated grids
**Submitting author:** @rhdtownsend (<a href="http://orcid.org/0000-0002-2522-8605">Richard Townsend</a>)
**Repository:** https://github.com/rhdtownsend/msg
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0
