### 2022-07-21 #4602 @AWehrhahn
Here are some comments on the points in the checklist that I didn't check off yet

Installation instructions: Bad
No python version given. Dependencies are given only by package name, no minimum versions requirement.
No automated installation of dependencies either.

Example usage: Good
    Clear easy to run example of the software. Although I would like to see this example included as a script in the github and not only in the documentation.

Functionality documentation: OK
    The API is documented with explanations for the parameters. But without example input/outputs.

Automated tests: Ok (but not automated)
    Only manual guide in the documentation
    automatic tests (in the test directory) are missing (where not installed?)

Community guidelines: Bad
    No guidelines given (at least not as far as I could find them). Best guess is to use the default GitHub features.

State of the field: Kind of?
    They state that other packages exists, that have a different main focus. But do not discuss
    how their performance differs from this package.
