### 2022-07-21 rhdtownsend/msg#9 @AWehrhahn
Compilation error
Following the instructions here: https://msg.readthedocs.io/en/latest/user-guide/quick-start.html#quick-start
the installation fails at the final compilation step. I have tracked it down to some misplaced files in src/forum.

Link to Review thread: https://github.com/openjournals/joss-reviews/issues/4602
