### 2022-07-21 rhdtownsend/msg#10 @AWehrhahn
Specify Python dependencies
Provide a list of python dependencies, with version numbers, ideally in a requirements.txt file. This will make installation much clearer.

Review thread: https://github.com/openjournals/joss-reviews/issues/4602
