### 2022-07-21 rhdtownsend/msg#11 @AWehrhahn
Make PyMSG pip installable
Currently there is no automated way to install the software package. This could be fixed by creating a pip compatible setup.py script and uploading the software to pypi.org.

Review Thread: https://github.com/openjournals/joss-reviews/issues/4602
