### 2022-07-21 rhdtownsend/msg#12 @AWehrhahn
Automated tests?
Hi,
I couldn't find any tests. There is nothing in my bin directory for the run_test.sh script to run?

Review Thread: https://github.com/openjournals/joss-reviews/issues/4602
