### 2022-07-21 rhdtownsend/msg#13 @AWehrhahn
Community Guidelines
Please give instructions on how to
1) Contribute to the software
2) Report issues or problems with the software
3) Seek support

These can also just be the github features (e.g. github issues)

Review thread: https://github.com/openjournals/joss-reviews/issues/4602
