### 2022-07-21 #4602 @AWehrhahn
Also please note that I have only tested the Python interface, not the C or Fortran APIs
