### 2022-07-25 #4602 @arfon
@ivastar - many thanks for agreeing to review this submission for JOSS!

Please take a look at the instructions above and let me know if you have any questions! https://github.com/openjournals/joss-reviews/issues/4602#issuecomment-1191517285
