### 2022-08-20 #4602 @arfon
> @arfon I've done the review. Do I write comments on the items I've left unchecked here?

Yes please! Also, if you've opened issues on the repo (https://github.com/rhdtownsend/msg) that are related to these unchecked items then please link to them here too.
