### 2022-08-22 rhdtownsend/msg#9 @ivastar
Following up on the review by @AWehrhahn, I was also unable to complete the installation via the instructions. The directory structure is not what the make file expects. I was able to install directly the current version from the repository. A possible solution would be to direct used to installing from a specific commit rather (e.g., [this example](https://stackoverflow.com/questions/13685920/install-specific-git-commit-with-pip)) rather than from a tar file.
