### 2022-08-22 rhdtownsend/msg#14 @ivastar
Performance claims for large files and additional library support
Claims are made for the handling of the graceful handling of large files. I was unable to verify these claims. Instructions on how to set the cache size and what is a reasonable value relative to the system memory would benefit users. Also not apparent what other libraries are supported and how those libraries can be converted to the format required by MSG.

Review thread: #https://github.com/openjournals/joss-reviews/issues/4602
