### 2022-08-22 #4602 @ivastar
Review completed. Most of my points agree with those of @AWehrhahn. I also only tested the Python interface. I added some comments to the issues that were already open and also added one more issue on the performance claims in handling large files.
