### 2022-11-06 rhdtownsend/msg#9 @rhdtownsend
It looks like there was a mistake in creating the tarball for the 1.0 release (goodness knows how that happened, I though I had checked it). I fixed this and reissued the tarball on 8/24/22.
