### 2022-11-18 560a67b
Added ascii_to_specint tool
