### 2022-11-19 a1d5215
Changed to ref-counted, memory-usage based caching caching for specgrids
