+++ src/include/cmsg.h
@@ -49,15 +49,15 @@ void get_specgrid_lam_min(SpecGrid specgrid, double *lam_min);
 void get_specgrid_lam_max(SpecGrid specgrid, double *lam_max);
 void get_specgrid_cache_lam_min(SpecGrid specgrid, double *cache_lam_min);
 void get_specgrid_cache_lam_max(SpecGrid specgrid, double *cache_lam_max);
-void get_specgrid_cache_count(SpecGrid specgrid, int *cache_count);
-void get_specgrid_cache_limit(SpecGrid specgrid, int *cache_limit);
+void get_specgrid_cache_usage(SpecGrid specgrid, long *cache_usage);
+void get_specgrid_cache_limit(SpecGrid specgrid, long *cache_limit);
 void get_specgrid_axis_x_min(SpecGrid specgrid, int i, double *axis_x_min);
 void get_specgrid_axis_x_max(SpecGrid specgrid, int i, double *axis_x_max);
 void get_specgrid_axis_label(SpecGrid specgrid, int i, char *axis_label);
 
 void set_specgrid_cache_lam_min(SpecGrid specgrid, double cache_lam_min, Stat *stat);
 void set_specgrid_cache_lam_max(SpecGrid specgrid, double cache_lam_max, Stat *stat);
-void set_specgrid_cache_limit(SpecGrid specgrid, int cache_limit, Stat *stat);
+void set_specgrid_cache_limit(SpecGrid specgrid, long cache_limit, Stat *stat);
 
 void interp_specgrid_intensity(SpecGrid specgrid, double x_vec[], double mu, int n, double lam[], double I[], Stat *stat, bool deriv_vec[]);
 void interp_specgrid_E_moment(SpecGrid specgrid, double x_vec[], int k, int n, double lam[], double E[], Stat *stat, bool deriv_vec[]);
