### 2022-11-20 de63740
Changed to ref-counted, memory-usage based caching caching for photgrids
