### 2022-11-20 cff9bc0
Changed cache fetch routines to return pointers, allowing proxy_specint_t to be retired
