+++ build/Makefile
@@ -65,7 +65,7 @@ libforum_OBJS = forum_m.o cinter_m.o const_m.o hdf5io_m.o kinds_m.o memory_m.o \
 
 libmsg_OBJS = msg_m.o axis_m.o vgrid_m.o ncubic_m.o \
               range_m.o range_sm.o comp_range_m.o lin_range_m.o log_range_m.o tab_range_m.o \
-              specgrid_m.o speccache_m.o specsource_m.o hdf5_specsource_m.o specint_m.o specint_sm.o limb_specint_m.o proxy_specint_m.o \
+              specgrid_m.o speccache_m.o specsource_m.o hdf5_specsource_m.o specint_m.o specint_sm.o limb_specint_m.o \
               photgrid_m.o photcache_m.o photsource_m.o hdf5_photsource_m.o spec_photsource_m.o mem_photsource_m.o photint_m.o photint_sm.o limb_photint_m.o \
               passband_m.o cubint_m.o stat_m.o file_m.o fit_m.o limb_m.o math_m.o
 libmsg_LIBS = libforum
