+++ src/cython/pycmsg.pyx
@@ -56,17 +56,19 @@ cdef extern from "cmsg.h":
     void get_specgrid_rank(void *specgrid, int *rank)
     void get_specgrid_lam_min(void *specgrid, double *lam_min)
     void get_specgrid_lam_max(void *specgrid, double *lam_max)
-    void get_specgrid_cache_usage(void *specgrid, long *cache_usage)
-    void get_specgrid_cache_limit(void *specgrid, long *cache_limit)
+    void get_specgrid_cache_usage(void *specgrid, int *cache_usage)
+    void get_specgrid_cache_limit(void *specgrid, int *cache_limit)
     void get_specgrid_cache_lam_min(void *specgrid, double *cache_lam_min)
     void get_specgrid_cache_lam_max(void *specgrid, double *cache_lam_max)
     void get_specgrid_axis_x_min(void *specgrid, int i, double *axis_x_min)
     void get_specgrid_axis_x_max(void *specgrid, int i, double *axis_x_max)
     void get_specgrid_axis_label(void *specgrid, int i, char *axis_label)
 
-    void set_specgrid_cache_limit(void *specgrid, long cache_limit, Stat *stat)
+    void set_specgrid_cache_limit(void *specgrid, int cache_limit, Stat *stat)
     void set_specgrid_cache_lam_min(void *specgrid, double cache_lam_min, Stat *stat)
     void set_specgrid_cache_lam_max(void *specgrid, double cache_lam_max, Stat *stat)
+
+    void flush_specgrid_cache(void *specgrid)
 
     void interp_specgrid_intensity(void *specgrid, double x_vec[], double mu,
                                    int n, double lam[], double I[], Stat *stat,
@@ -86,13 +88,15 @@ cdef extern from "cmsg.h":
     void unload_photgrid(void *photgrid)
 
     void get_photgrid_rank(void *photgrid, int *rank)
-    void get_photgrid_cache_usage(void *photgrid, long *cache_usage)
-    void get_photgrid_cache_limit(void *photgrid, long *cache_limit)
+    void get_photgrid_cache_usage(void *photgrid, int *cache_usage)
+    void get_photgrid_cache_limit(void *photgrid, int *cache_limit)
     void get_photgrid_axis_x_min(void *photgrid, int i, double *axis_x_min)
     void get_photgrid_axis_x_max(void *photgrid, int i, double *axis_x_max)
     void get_photgrid_axis_label(void *photgrid, int i, char *axis_label)
 
-    void set_photgrid_cache_limit(void *photgrid, long cache_limit, Stat *stat)
+    void set_photgrid_cache_limit(void *photgrid, int cache_limit, Stat *stat)
+
+    void flush_photgrid_cache(void *photgrid)
 
     void interp_photgrid_intensity(void *photgrid, double x_vec[], double mu,
                                    double *I, Stat *stat, bool deriv_vec[])
@@ -153,7 +157,7 @@ def _get_specgrid_lam_max(uintptr_t specgrid):
 
 def _get_specgrid_cache_usage(uintptr_t specgrid):
 
-    cdef long cache_usage
+    cdef int cache_usage
 
     get_specgrid_cache_usage(<void *>specgrid, &cache_usage)
 
@@ -162,7 +166,7 @@ def _get_specgrid_cache_usage(uintptr_t specgrid):
 
 def _get_specgrid_cache_limit(uintptr_t specgrid):
 
-    cdef long cache_limit
+    cdef int cache_limit
     
[…]
