### 2022-11-22 a8c9496
Added flush_cache method
