+++ python/pymsg/photgrid.py
@@ -114,18 +114,23 @@ class PhotGrid:
     
     @property
     def cache_usage(self):
-        """int: Current memory usage (in bytes) of data in grid cache."""
+        """int: Current memory usage (in kilobytes) of data in grid cache."""
         return pyc._get_photgrid_cache_usage(self._photgrid)
 
 
     @property
     def cache_limit(self):
-        """int: Maximum memory usage (in bytes) of data in grid cache. Set to 0 to disable 
+        """int: Maximum memory usage (in kilobytes) of data in grid cache. Set to 0 to disable
            caching."""
         return pyc._get_photgrid_cache_limit(self._photgrid)
     @cache_limit.setter
     def cache_limit(self, cache_limit):
         pyc._set_photgrid_cache_limit(self._photgrid, limit)
+
+
+    def flush_cache(self):
+        """Flush the grid cache"""
+        pyc._flush_photgrid_cache(self._photgrid)
 
 
     def intensity(self, x, mu, deriv=None):
