+++ docs/source/ref-guide/data-schema.rst
@@ -1,0 +1,23 @@
+.. _data-schema:
+
+***********
+Data Schema
+***********
+
+MSG spectroscopic and photometric grid data are stored on disk in
+`HDF5 <https://www.hdfgroup.org/solutions/hdf5/>`__ files with a
+bespoke schema. Throughout this documentation, these files are known
+as `specgrid` and `photgrid` files, respectively.
+
+.. toctree::
+   :maxdepth: 2
+
+   data-schema/axis.rst
+   data-schema/limb.rst
+   data-schema/range.rst
+   data-schema/photgrid.rst
+   data-schema/photint.rst
+   data-schema/specgrid.rst
+   data-schema/specint.rst
+   data-schema/vgrid.rst
+   
