+++ docs/source/ref-guide/fortran-interface/params.rst
@@ -1,5 +1,7 @@
 .. _fortran-params:
 
+ .. f:currentmodule:: fmsg_m
+
 Parameters
 ----------
 
@@ -9,85 +11,91 @@ Parameters
    Length of :f:type:`axis_t` labels.
 
 
-.. f:variable:: int STAT_OK
+.. f:variable:: int ~fmsg_m/STAT_OK
    :attrs: parameter
 
-   Status code indicating call completed without error.
+   Status code indicating procedure call completed without error.
 
 
 .. f:variable:: int STAT_OUT_OF_BOUNDS_RANGE_LO
    :attrs: parameter
    
-   Status code indicating call encountered an out-of-bounds reference, below the range minimum.
+   Status code indicating procedure call encountered an out-of-bounds reference, below the range minimum.
 
 
 .. f:variable:: int STAT_OUT_OF_BOUNDS_RANGE_HI
    :attrs: parameter
    
-   Status code indicating call encountered an out-of-bounds reference, above the range maximum.
+   Status code indicating procedure call encountered an out-of-bounds reference, above the range maximum.
    
 
 .. f:variable:: int STAT_OUT_OF_BOUNDS_AXIS_LO
    :attrs: parameter
    
-   Status code indicating call encountered an out-of-bounds reference, below the axis minimum.
+   Status code indicating procedure call encountered an out-of-bounds reference, below the axis minimum.
 
 
 .. f:variable:: int STAT_OUT_OF_BOUNDS_AXIS_HI
    :attrs: parameter
    
-   Status code indicating call encountered an out-of-bounds reference, above the axis maximum.
+   Status code indicating procedure call encountered an out-of-bounds reference, above the axis maximum.
    
 
 .. f:variable:: int STAT_OUT_OF_BOUNDS_LAM_LO
    :attrs: parameter
    
-   Status code indicating call encountered an out-of-bounds reference, below the wavelength minimum.
+   Status code indicating procedure call encountered an out-of-bounds reference, below the wavelength minimum.
 
 
 .. f:variable:: int STAT_OUT_OF_BOUNDS_LAM_HI
    :attrs: parameter
    
-   Status code indicating call encountered an out-of-bounds reference, above the wavelength maximum.
+   Status code indicating procedure call encountered an out-of-bounds reference, above the wavelength maximum.
    
 
 .. f:variable:: int STAT_OUT_OF_BOUNDS_MU_LO
    :attrs: parameter
    
-   Status code indicating call encountered an out-of-bounds reference, below the emergence cosine minimum.
+   Status code indicating procedure call encountered an out-of-bounds reference, below the emergence cosine minimum.
 
 
 .. f:variable:: int STAT_OUT_OF_BOUNDS_MU_HI
    :attrs: parameter
    
-   Status code indicating call encountered an out-of-bounds reference, above the emergence cosine maximum.
+   Status code indicating procedure call encountered an out-of-bounds reference, above the emergence cosine maximum.
    
 
 .. f:variable:: int STAT_UNAVAILABLE_DATA
    :attrs: parameter
    
-   Status code indicating call encountered unavailable data.
+   Status code indicating procedure call encountered unavailable data.
    
 
 .. f:variable:: int STAT_INVALID_ARGUMENT
    :attrs: parameter
    
-   Status code indicating call encountered an invalid argument.
+   Status code indicating procedure call encountered an invalid argument.
 
 
 .. f:variable:: int STAT_FILE_NOT_FOUND
    :attrs: parameter
    
-   Status code indicating call encountered a file that could not be found.
+   Status code indicating procedure call encountered a file that could not be found.
 
 
 .. f:variable:: int STAT_INVALID_FILE_TYPE
    :attrs: parameter
    
[…]
