### 2022-11-26 9e0000c
Fixes for issue #15
