### 2022-11-26 267698b
Fixes for issue #16
