### 2022-11-27 #4602 @arfon
:wave: @rhdtownsend – it's now three months since we heard from you on this submission. Could you re-confirm your interest in pursing this publication in JOSS?
