### 2022-11-27 #4602 @rhdtownsend
Yes, I am still interested — been working on changes to address the referees' remarks.

cheers,

Rich

> On Nov 27, 2022, at 6:29 AM, Arfon Smith ***@***.***> wrote:
>
>
> 👋 @rhdtownsend – it's now three months since we heard from you on this submission. Could you re-confirm your interest in pursing this publication in JOSS?
