### 2022-11-27 6383104
Updated testing infrastructure, to provide summary output
