+++ test/run_tests.sh
@@ -1,5 +1,12 @@
 #!/bin/sh
 
-for t in ../bin/utest_*; do
-    $t
+rm -f test.out
+
+for t in ${BINDIR}/utest_*; do
+    $t | tee -a test.out
 done
+
+echo "==============="
+echo " Passes:   " `grep -c PASS test.out`
+echo " Failures: " `grep -c FAIL test.out`
+echo "==============="
