+++ test/Makefile
@@ -1,0 +1,12 @@
+# File     : Makefile
+# Purpose  : makefile for test subdir
+
+# Rules
+
+test :
+	@./run_tests.sh
+
+clean almostclean :
+	rm -f test.out
+
+.PHONY: test
