+++ docs/source/ref-guide/data-schema/passband-file.rst
@@ -1,0 +1,17 @@
+.. _data-schema-passband-file:
+
+Passband File
+=============
+
+.. list-table::
+   :widths: 15 10 10 65
+   :header-rows: 1
+
+   * - Item
+     - Objtype
+     - Datatype
+     - Definition
+   * - :code:`/`
+     - group
+     - :ref:`data-schema-passband-group`
+     - passband.
