+++ docs/source/ref-guide/data-schema/limb-group.rst
@@ -1,0 +1,25 @@
+.. _data-schema-limb-group:
+
+Limb Group
+==========
+
+.. list-table::
+   :widths: 15 10 10 65
+   :header-rows: 1
+
+   * - Item
+     - Objtype
+     - Datatype
+     - Definition
+   * - :code:`TYPE`
+     - attribute
+     - character
+     - literal :code:`'limb_t'`.
+   * - :code:`REVISION`
+     - attribute
+     - integer
+     - literal :code:`1`.
+   * - :code:`law`
+     - attribute
+     - integer
+     - limb-darkening law code.
