### 2022-11-27 5bb0d93
Chenged test and tool build to on by default, & updated targets
