+++ docs/source/ref-guide/c-interface.rst
@@ -8,7 +8,7 @@ The C interface is provided through a set of functions defined in the
 :file:`cmsg.h` header file. These functions are implemented using the
 `C interoperability capabilities
 <https://gcc.gnu.org/onlinedocs/gfortran/Interoperability-with-C.html>`__
-in Fortram 2003/2008/2018, and are based around a pair of `void *`
+in Fortran 2003/2008/2018, and are based around a pair of `void *`
 typedefs --- :c:type:`SpecGrid` and :c:type:`PhotGrid` --- that store
 pointers to corresponding Fortran :f:type:`specgrid_t` and
 :f:type:`photgrid_t` objects.
