+++ docs/source/user-guide/contributing.rst
@@ -8,7 +8,7 @@ Code
 To contribute code toward the MSG project, from bug fixes through to
 major new features, the preferred method is the standard
 fork---branch---pull-request paradigm described `here
-<https://git-scm.com/book/en/v2/GitHub-Contributing-to-a-Project>'. If
+<https://git-scm.com/book/en/v2/GitHub-Contributing-to-a-Project>`__. If
 your pull request languishes for more than a couple of weeks without
 receiving a response, consider opening an issue to give the developers
 a nudge.
@@ -24,8 +24,8 @@ Data
 ====
 
 To contribute data toward the project (in the form of spectral
-intensity or flux grids), the first step is to create a set of filest convert these data into
-MSG's HDF5-based `data schema`.
+intensity or flux grids), the first step is to create a set of files
+convert these data into MSG's HDF5-based `data schema`.
 
 
 
