+++ docs/source/user-guide/walkthroughs.rst
@@ -1,0 +1,14 @@
+.. _walkthroughs:
+
+************
+Walkthroughs
+************
+
+The best way to familiarize yourself with MSG is use it.
+
+.. toctree::
+   :maxdepth: 2
+
+   walkthroughs/python-walkthrough.rst
+   walkthroughs/fortran-walkthrough.rst
+   walkthroughs/c-walkthrough.rst
