### 2022-12-04 da3b716
Changed Python stuff to be installable via pip
