+++ python/README.rst
@@ -1,0 +1,22 @@
+*****
+PyMSG
+*****
+
+Overview
+========
+
+PyMSG provides a Python interface to MSG (Multidimensional Stellar
+Grids), a library that synthesizes astrophysical spectra and
+photometric colors via interpolation in pre-calculated grids.  Given a
+set of stellar atmospheric parameters (e.g., effective temperature,
+gravity and metallicity), MSG can calculate the specific intensity of
+the emergent radiation field as a function of wavelength and angle. It
+can also convolve these data with appropriate filters and response
+functions, to evaluate colors in a wide variety of photometric
+systems.
+
+Documentation
+=============
+
+Full documentation for MSG (including PyMSG) can be found at
+https://msg.readthedocs.io/en/latest.
