+++ docs/source/ref-guide/data-schema/files.rst
@@ -1,0 +1,13 @@
+.. _data-schema-files:
+
+Files
+=====
+
+.. toctree::
+   :maxdepth: 2
+
+   files/specint.rst
+   files/specgrid.rst
+   files/passband.rst
+   files/photgrid.rst
+   files/photint.rst
