+++ docs/source/ref-guide/data-schema/groups.rst
@@ -1,0 +1,18 @@
+.. _data-schema-groups:
+
+Groups
+======
+
+.. toctree::
+   :maxdepth: 2
+
+   groups/axis.rst
+   groups/cubint.rst
+   groups/limb.rst
+   groups/range.rst
+   groups/passband.rst
+   groups/photgrid.rst
+   groups/photint.rst
+   groups/specgrid.rst
+   groups/specint.rst
+   groups/vgrid.rst

[…]
