+++ docs/source/ref-guide/python-interface.rst
@@ -4,10 +4,8 @@
 Python Interface
 ****************
 
-The Python interface is provided through the :py:mod:`pymsg` module,
-which defines the :py:class:`pymsg.SpecGrid` and
-:py:class:`pymsg.PhotGrid` classes. It is implemented by `Cython
-<https://cython.org/>`__-wrapped calls to the :ref:`c-interface`.
+The Python interface is provided through the :py:mod:`pymsg` module.
+
 
 API Specification
 =================
@@ -15,14 +13,4 @@ API Specification
 .. toctree::
    :maxdepth: 2
 
-   python-interface/specgrid.rst
-   python-interface/photgrid.rst
-
-Importing
-=========
-
-The :py:mod:`pymsg` module is provided in the :file:`$MSG_DIR/python`
-directory. To import the module, this directory should be added to the
-Python path --- for instance by setting the :envvar:`PYTHONPATH`
-environment variable, as discussed `here <https://docs.python.org/3/using/cmdline.html#envvar-PYTHONPATH>`__.
-
+   python-interface/classes.rst
