+++ docs/source/ref-guide/data-schema/groups/range.rst
@@ -1,125 +1,0 @@
-.. _data-schema-groups-range:
-
-:grouptype:`range`
-==================
-
-Linear
-------
-
-.. list-table::
-   :widths: 15 10 10 65
-   :header-rows: 1
-
-   * - Item
-     - Objtype
-     - Datatype
-     - Definition
-   * - :code:`TYPE`
-     - attribute
-     - character
-     - literal :code:`'lin_range_t'`.
-   * - :code:`REVISION`
-     - attribute
-     - integer
-     - literal :code:`1`.
-   * - :code:`x_0`
-     - attribute
-     - real
-     - abscissa start value.
-   * - :code:`dx`
-     - attribute
-     - real
-     - abscissa spacing.
-   * - :code:`n`
-     - attribute
-     - integer
-     - number of points.
-
-
-Logarithmic
------------
-
-.. list-table::
-   :widths: 15 10 10 65
-   :header-rows: 1
-
-   * - Item
-     - Objtype
-     - Datatype
-     - Definition
-   * - :code:`TYPE`
-     - attribute
-     - character
-     - literal :code:`'log_range_t'`.
-   * - :code:`REVISION`
-     - attribute
-     - integer
-     - literal :code:`1`.
-   * - :code:`logx_0`
-     - attribute
-     - real
-     - logarithm of abscissa start value.
-   * - :code:`dlogx`
-     - attribute
-     - real
-     - abscissa logarithmic spacing.
-   * - :code:`n`
-     - attribute
-     - integer
-     - number of points.
-     
-
-Tabulated
----------
-
-.. list-table::
-   :widths: 15 10 10 65
-   :header-rows: 1
-
-   * - Item
-     - Objtype
-     - Datatype
-     - Definition
-   * - :code:`TYPE`
-     - attribute
-     - character
-     - literal :code:`'tab_range_t'`.
-   * - :code:`REVISION`
-     - attribute
-     - integer
-     - literal :code:`1`.
-   * - :code:`x`
-     - dataset
-     - real(:)
-     - abscissa values.
-
-
-Composite
----------
-
-.. list-table::
-   :widths: 15 10 10 65
-   :header-rows: 1
-
-   * - Item
-     - Objtype
-     - Datatype
-     - Definition
-   * - :code:`TYPE`
-     - attribute
-     - character
-     - literal :code:`'comp_range_t'`.
-   * - :code:`REVISION`
-     - attribute
-     - integer
-     - literal :code:`1`.
-   * - :code:`n_ranges`
-     - dataset
-     - integer
-     - number of sub-ranges.
-   * - :code:`ranges[i]`
-     - group
-     - :groupref:`range`
-     - sub-ranges (:code:`i = 1, ..., n_ranges`).
-
-

[…]
