+++ docs/source/user-guide/contributing.rst
@@ -5,51 +5,28 @@ Contributing
 Code
 ====
 
-To contribute code toward the MSG project, from bug fixes through to
-major new features, the preferred method is the standard
+To contribute code to the MSG project, from bug fixes through to major
+new features, the preferred method is the standard
 fork---branch---pull-request paradigm described `here
 <https://git-scm.com/book/en/v2/GitHub-Contributing-to-a-Project>`__. If
 your pull request languishes for more than a couple of weeks without
-receiving a response, consider opening an issue to give the developers
-a nudge.
+receiving a response, consider :ref:`opening an issue <open-an-issue>`
+to give the developers a nudge.
 
 Documentation
 =============
 
-To contribute documentaion toward the project, follow the same
+To contribute documentation toward the project, follow the same
 approach as above for code. (The ReStructured Text source files for
-the documentation reside in the :file:`docs/source` subdirectory.)
+the documentation reside in the :file:`docs/source` sub-directory.)
 
 Data
 ====
 
 To contribute data toward the project (in the form of spectral
-intensity or flux grids), a number of options are available depending
-on your level of technical expertise. Both involve first converting
-your data into HDF5 files following the :ref:`specint
-<data-schema-specint-files>` data schemas.
-
-Using the Tools
----------------
-
-The easier but less flexible approach is to use one of the :ref:`tools
-provided` to convert spectra from a 
-
-
-Building your Own Tools
------------------------
-
-
-
-the first step
-is to create a set of files convert these data into MSG's HDF5-based
-`data schema`.
-
-
-
-
-
-
-
-
-
+intensity or flux grids), a number of options are available. The
+:ref:`data-files-importing` section explains how to import an existing
+grid; but if this lies beyond your technical expertise, then the MSG
+team will be happy to assist. Either way, if you want to make your
+data publicly available and listed in the :ref:`grid-files` chapter,
+then please :ref:`open an issue <open-an-issue>`.
