### 2022-12-06 rhdtownsend/msg#14 @rhdtownsend
Details on cache function, together with a demonstration of caching in action, are provided in a new Data Caching section in the documentation:

https://msg.readthedocs.io/en/latest/user-guide/data-caching.html

Likewise, a discussion of importing grids from other sources is provided in a new Data Files section:

https://msg.readthedocs.io/en/latest/user-guide/data-files.html
