### 2022-12-06 rhdtownsend/msg#11 @rhdtownsend
pymsg is now pip-installable, as described in the updated Quick Start and Installation documentation:

https://msg.readthedocs.io/en/latest/user-guide/quick-start.html
file:///Users/townsend/devel/msg/docs/build/html/ref-guide/installation.html

For now, I dont think it makes sense to upload pymsg to the Python Package Index. The module can't function without the MSG libraries, and therefore a user will always have to install MSG as well. It seems wise to therefore ship pymsg with MSG.
