### 2022-12-06 8ad85ed
Merged individual submodules into main module, added docstring
