+++ python/src/pymsg/__init__.py
@@ -15,5 +15,500 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from .specgrid import SpecGrid
-from .photgrid import PhotGrid
+"""This module provides the Python interface to the Multidimensional
+Specgral Grids (MSG) package.
+"""
+
+import os
+import sys
+import numpy as np
+
+# Set path & import pycmsg
+
+sys.path.insert(0, os.path.join(os.environ['MSG_DIR'], 'lib'))
+import pycmsg as pyc
+
+# Class definitions
+
+class SpecGrid:
+    r"""The SpecGrid class represents a grid of spectroscopic intensity data.
+
+    This grid may be used to interpolate the intensity (or related
+    quantities) across a wavelength abscissa and for a set of
+    atmosphere parameter values.
+
+    """
+
+    def __init__(self, file_name):
+        """SpecGrid constructor (via loading data from a `specgrid` file).
+
+        Args:
+            file_name (string): Name of the file
+
+        Returns:
+            pymsg.SpecGrid: Constructed object.
+
+        Raises:
+            FileNotFound: If the the file cannot be found.
+            TypeError: If the file contains an incorrect datatype.
+        """
+
+        self._specgrid = pyc._load_specgrid(file_name)
+
+        self._rank = pyc._get_specgrid_rank(self._specgrid)
+
+        self._axis_labels = []
+        self._axis_x_min = {}
+        self._axis_x_max = {}
+
+        for i in range(self._rank):
+
+            axis_label = pyc._get_specgrid_axis_label(self._specgrid, i)
+
+            self._axis_labels += [axis_label]
+            self._axis_x_min[axis_label] = pyc._get_specgrid_axis_x_min(self._specgrid, i)
+            self._axis_x_max[axis_label] = pyc._get_specgrid_axis_x_max(self._specgrid, i)
+
+        
+    def __dealloc__(self):
+
+        pyc._unload_specgrid(self._specgrid)
+
+        self._specgrid = None
+
+
+    def _vector_args(self, x, deriv):
+
+        x_vec = np.array([x[key] for key in self._axis_labels])
+
+        if deriv is not None:
+            deriv_vec = np.array([key in deriv for key in self._axis_labels],
+                                 dtype=np.uint8)
+        else:
+            deriv_vec = np.array([False]*self.rank, dtype=np.uint8)
+
+        return x_vec, deriv_vec
+
+    
+    @property
+    def rank(self):
+        """int: Number of dimensions in grid."""
+        return self._rank
+
+    
+    @property
+    def axis_labels(self):
+        """list: Atmosphere parameter axis labels."""
+        return self._axis_labels
+
+    
+    @property
+    def axis_x_min(self):
+        """dict: Atmosphere parameter axis minima."""
+        return self._axis_x_min
+
+    
+    @property
+    def axis_x_max(self):
+        """dict: Atmosphere parameter axis maxima."""
+        return self._axis_x_max
+
+    
+    @property
+    def lam_min(self):
+        """double: Minimum wavelength of grid (Å)."""
+        return pyc._get_specgrid_lam_min(self._specgrid)
+
+    
+    @property
+    def lam_max(self):
+        """double: Maximum wavelength of grid (Å)."""
+        return pyc._get_specgrid_cache_lam_max(self._specgrid)
+
+
+    @property
+    def cache_lam_min(self):
+        """double: Minimum wavelength of grid cache (Å)."""
+        return pyc._get_specgrid_cache_lam_min(self._specgrid)
+    @cache_lam_min.setter
+    def cache_lam_min(self, cache_lam_min):
[…]

[…]
