+++ python/pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "pymsg"
-version = "1.1rc1"
+version = "1.1rc2"
 authors = [
   { name="Rich Townsend", email="rhtownsend@wisc.edu" },
 ]
