### 2022-12-07 11ec846
Added doc for ascii_to_specint
