+++ docs/source/ref-guide/python-interface.rst
@@ -10,6 +10,8 @@ The Python interface is provided through the :py:mod:`pymsg` module.
 API Specification
 =================
 
+.. py:module:: pymsg
+
 .. toctree::
    :maxdepth: 2
 
