+++ docs/source/ref-guide/data-schema/groups.rst
@@ -7,12 +7,12 @@ Groups
 .. include:: groups/comp_range.inc
 .. include:: groups/cubint.inc
 .. include:: groups/limb.inc
+.. include:: groups/limb_photint.inc
+.. include:: groups/limb_specint.inc
 .. include:: groups/lin_range.inc
 .. include:: groups/log_range.inc
 .. include:: groups/passband.inc
 .. include:: groups/photgrid.inc
-.. include:: groups/photint.inc
 .. include:: groups/specgrid.inc
-.. include:: groups/specint.inc
 .. include:: groups/tab_range.inc
 .. include:: groups/vgrid.inc
