### 2022-12-10 2a69631
Fixed docstrings
