+++ paper.bib
@@ -37,6 +37,19 @@
 	year = 2018,
 	doi = {10.1051/0004-6361/201732147}}
 
+@article{Czekala:2015,
+        author       = {{Czekala}, Ian and {Andrews}, Sean M. and {Mandel}, Kaisey S. and
+                        {Hogg}, David W. and {Green}, Gregory M.},
+        title        = "{Constructing a Flexible Likelihood Function for Spectroscopic Inference}",
+        journal      = {\apj},
+        year         = "2015",
+        month        = "Oct",
+        volume       = {812},
+        number       = {2},
+        pages        = {128},
+        doi          = {10.1088/0004-637X/812/2/128},
+}
+
 @article{de-Laverny:2012,
 	author = {{de Laverny}, P. and {Recio-Blanco}, A. and {Worley}, C.~C. and {Plez}, B.},
 	journal = {Astronomy \& Astrophysics},
@@ -88,6 +101,16 @@
 	year = 2007,
 	doi = {10.1086/511270}}
 
+@article{McIlroy:1978,
+         author = {{McIlroy}, M. D. and {Pinson}, E. N. and {Tague}, B.A.},
+	 title = {UNIX Time-Sharing System: Forward},
+	 journal = {Bell System Technical Journal},
+	 volume = {57},
+	 number = {6},
+	 pages = {1899},
+	 month = aug,
+	 year = 1978}
+
 @article{Meszaros:2013,
 	author = {{M{\'e}sz{\'a}ros}, Sz. and {Allende Prieto}, C.},
 	journal = {Monthly Notices of the Royal Astronomical Society},
@@ -98,6 +121,18 @@
 	volume = {430},
 	year = 2013,
 	doi = {10.1093/mnras/stt130}}
+
+@misc{STScI:2020,
+       author = {{STScI Development Team}},
+        title = "{stsynphot: synphot for HST and JWST}",
+ howpublished = {Astrophysics Source Code Library, record ascl:2010.003},
+         year = 2020,
+        month = oct,
+          eid = {ascl:2010.003},
+        pages = {ascl:2010.003},
+archivePrefix = {ascl},
+       eprint = {2010.003}
+}
 
 @article{Zsargo:2020,
 	author = {{Zsarg{\'o}}, J. and {Fierro-Santill{\'a}n}, C.~R. and {Klapp}, J. and {Arrieta}, A. and {Arias}, L. and {Valencia}, J.~M. and {Sigalotti}, L. Di G. and {Hareter}, M. and {Puebla}, R.~E.},
