+++ paper.md
@@ -31,7 +31,7 @@ While the spectrum of the light emitted by a star can be calculated by
 simulating the flow of radiation through each layer of the star's
 atmosphere, this process is computationally expensive.  Therefore, it
 is often far more efficient to pre-calculate spectra over a grid of
-atmospheric parameters, and then interpolate within this grid.  `MSG`
+photospheric parameters, and then interpolate within this grid.  `MSG`
 (short for Multidimensional Spectral Grids) is a software package that
 implements this interpolation capability.
 
@@ -42,29 +42,44 @@ astronomical literature --- examples include @Lanz:2003, @Lanz:2007,
 @Kirby:2011, @de-Laverny:2012, @Husser:2013, @Allende-Prieto:2018,
 @Chiavassa:2018 and @Zsargo:2020. However, the ecosystem of software
 packages that offer users the ability to interpolate in these grids is
-much more limited. iSPEC [@Blanco-Cuaresma:2014] and FERRE
-[@Allende-Prieto:2015] stand out in this category, but their principal
-focus is spectral analysis (determining stellar atmosphere parameters
-by fitting observed spectra) rather than interpolation; and, as
-monolithic packages they are not well suited to modular incorporation
-within other projects. These considerations motivate us to develop
-MSG.
+much more limited:
+
+* FERRE [@Allende-Prieto:2015] supports piecewise-cubic interpolation
+  in an arbitrary number of photospheric parameters, but is restricted
+  to grids with rectilinear boundaries. Moreover, as a monolithic
+  executable it is not well suited to modular embedding within other
+  projects.
+
+* Starfish [@Czekala:2015] offers a Python API supporting
+  piecewise-linear interpolation in an arbitrary number of
+  photospheric parameters [see @Meszaros:2013 for a discussion of
+  the limitations of linear schemes].
+
+* stsynphot [@STScI:2020] also offers a Python API supporting
+  piecewise-linear interpolation, but is restricted to three
+  photospheric parameters and a hard-coded selection of grids.
+
+The limitations of these packages stem in part from their purpose:
+each has a broader focus than spectral interpolation alone. Guided by
+the Unix philosophy of 'make each program do one thing well'
+[@McIlroy:1978], this motivates us to develop MSG.
 
 # Capabilities
 
 MSG is implemented as a software library with Python, Fortran 2008 and
-C bindings. These APIs are underpinned by OpenMP-parallelized Fortran
-code that performs energy-conservative interpolation in wavelength
-$\lambda$, parametric interpolation in direction cosine $\mu$ using
-limb-darkening laws, and $C^{1}$-continuous cubic tensor-product
-interpolation in an arbitrary number of atmosphere parameters
-(effective temperature $T_{\mathrm{eff}}$, surface gravity $g$,
-metallicity [Fe/H], etc.). Although the topology of grid points must
-remain Cartesian, their distribution along each separate dimension
-need not be uniform. Attempts to interpolate in regions with missing
-data (e.g., ragged grid boundaries and/or holes) are handled
-gracefully via exceptions (Python) or returned status codes (Fortran
-and C).
+C bindings. These APIs each provide routines for interpolating
+specific intensity and flux spectra. They are underpinned by
+OpenMP-parallelized Fortran code that performs energy-conservative
+interpolation in wavelength $\lambda$, parametric interpolation in
+direction cosine $\mu$ using limb-darkening laws, and
+$C^{1}$-continuous cubic tensor-product interpolation in an arbitrary
+number of photospheric parameters (effective temperature
+$T_{\mathrm{eff}}$, surface gravity $g$, metallicity [Fe/H],
+etc.). Although the topology of grid points must remain Cartesian,
+their distribution along each separate dimension need not be
+uniform. Attempts to interpolate in regions with missing data (e.g.,
[…]
