+++ docs/source/user-guide/exception-handling.rst
@@ -11,10 +11,10 @@ This chapter discusses how MSG handles exceptions. These can arise in a variety 
 * Attempts to load grids from files that are missing or corrupt.
 * Attempts to interpolate at locations outside the bounds of a grid.
 * Attempts to interpolate at locations where grid data are missing (so-called *grid voids*).
-* Attempts to interpolate with incomplete specification of atmosphere parameters.
+* Attempts to interpolate with incomplete specification of photospheric parameters.
 * Attempts to interpolate for invalid wavelength and/or angle parameters.
 
-When an exception occurs, how it is signaled depends on the language interface being used.
+When an exception occurs, how it's signaled depends on the language interface being used.
 
 Python
 ======
@@ -28,11 +28,11 @@ Fortran
 =======
 
 Using the Fortran interface, exceptions are signaled through the
-optional procedure argument `stat`. If this argument is
+optional procedure argument :code:`stat`. If this argument is
 present, then on return it is set to one of the status code values
 defined in the :ref:`Fortran parameters <fortran-params>` section. The
 value :f:var:`STAT_OK` indicates that no problem was encountered;
-other values signal an exception. If `stat` is not present when
+other values signal an exception. If :code:`stat` is not present when
 an exception occurs, then execution halts with an error message
 printed to standard output.
 
@@ -40,10 +40,10 @@ C
 =
 
 Using the C interface, exceptions are signaled through the pointer
-function argument `stat`. If this pointer is non-null, then on
-return the pointer target `*stat` is set to one of the status code values
+function argument :code:`stat`. If this pointer is non-null, then on
+return the pointer target :code:`*stat` is set to one of the status code values
 defined in the :ref:`C enums <c-enums>` section. The value
 :c:enumerator:`STAT_OK` indicates that no problem was encountered; other
-values signal an exception. If `stat` is null when an exception
+values signal an exception. If :code:`stat` is null when an exception
 occurs, then execution halts with an error message printed to standard
 output.
