+++ docs/source/ref-guide/c-interface/funcs.rst
@@ -3,8 +3,17 @@
 Functions
 ---------
 
+SpecGrid
+~~~~~~~~
+
 .. include:: specgrid.inc
 
-----
+PhotGrid
+~~~~~~~~
 
 .. include:: photgrid.inc
+
+Other
+~~~~~
+
+.. include:: version.inc
