+++ python/pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "pymsg"
-version = "1.1rc2"
+version = "1.1"
 authors = [
   { name="Rich Townsend", email="rhtownsend@wisc.edu" },
 ]
