### 2022-12-11 a63c344
Update README.rst
Fixed doc link to point to stable.
