+++ README.rst
@@ -13,7 +13,7 @@ The MSG Library
    :target: https://github.com/rhdtownsend/msg/issues
 .. image:: https://img.shields.io/readthedocs/msg
    :alt: Read the Docs
-   :target: https://msg.readthedocs.io/en/latest
+   :target: https://msg.readthedocs.io/en/stable
 .. image:: https://joss.theoj.org/papers/fdb605600d48545f98dfb9f825dc833d/status.svg
    :alt: Journal of Open Source Software
    :target: https://joss.theoj.org/papers/fdb605600d48545f98dfb9f825dc833d
