### 2022-12-11 #4602 @rhdtownsend
I'm ready for my changes to MSG to be reveiwed. In brief, here's how I've modified the project in response to the comments from @AWehrhahn:

* **Installation**: a mistake in generating the tar.gz source code file has been fixed, so installation should now proceed without issue
* **Installation Instructions**: the Python dependencies are now listed in the `pyproject.toml` file (and the pymsg module can be installed using `pip`, which of course handles automatic dependency checking)
* **Example Usage**: the Python walkthrough notebook is available in GitHub (docs/source/user-guide/walkthroughs/python-walkthrough.ipynb), and a note mentioning this is included in the docs.
* **Automated Tests**: the tests can now be run with `make test` (and the docs have been updated to recommend this). The functionality run during testing has been broadened, and at the end of `make test` the number of passes and failures is summarized.
* **Community Guidelines**: the docs now include a [Contributing](https://msg.readthedocs.io/en/stable/user-guide/contributing.html) chapter that discusses how to engage with the project.
* **State of the Field**: the paper discusses similar packages in a little more detail, highlighting where each falls somewhat short of being an ideal solution.

And here's how I've modified the project in response to the comments from @ivastar:

* **Installation**: see above.
* **Performance Claims**: the docs now include a [Data Caching](https://msg.readthedocs.io/en/stable/user-guide/data-caching.html) chapter that demonstrates how limits can be imposed on memory usage, and how performance is impacted by caching. (And the caching API now works in terms of MB used, rather than the less-intuitive number of grid nodes used).
* **Installation Instructions**: see above.
* **Automated Tests**: see above.
* **Community Guidelines**: see above.

I also bumped the version number to 1.1, to reflect the changes I made to the caching API.

Thanks to both reviewers for their helpful remarks!
