### 2022-12-20 #4602 @rhdtownsend
Hi Arfon —

I committed my changes, and submitted responses to the referee’s remarks, last week. Is there anything else I need to do to effect a ‘resubmit’ of the paper?

cheers,

Rich

> On Dec 9, 2022, at 9:07 AM, RICHARD H D TOWNSEND ***@***.***> wrote:
>
> I hope to be submitting in a week or so — just putting some finishing touches in place.
>
>> On Dec 9, 2022, at 4:04 AM, Arfon Smith ***@***.***> wrote:
>>
>>
>> Thanks for getting back to us. Do you have an estimate of when you might wrap up the changes?
>> —
>> Reply to this email directly, view it on GitHub, or unsubscribe.
>> You are receiving this because you were mentioned.Message ID: ***@***.***>
>
