### 2022-12-23 #4602 @ivastar
@rhdtownsend thank you for making all these excellent changes!

I ran through the installation process again and it worked for me. One minor comments: the MESA Software Development Kit link in the documentation is incorrect (https://msg.readthedocs.io/en/stable/ref-guide/mesa-sdk)

I'm ready to sign off.
