### 2022-12-23 c11fa58
Update README.rst
Fixed doc link in text.
