### 2023-01-17 #4602 @rhdtownsend
Hi folks —

Following up on this — any progress on the review?

cheers,

Rich

> On Dec 23, 2022, at 11:34 AM, Ansgar Wehrhahn ***@***.***> wrote:
>
>
> Hi, im on christmas Break right now, will look at it I'm January. Happy Holidays 😊
> ________________________________
> From: Rich Townsend ***@***.***>
> Sent: Friday, December 23, 2022 5:11:26 PM
> To: openjournals/joss-reviews ***@***.***>
> Cc: Ansgar Wehrhahn ***@***.***>; Mention ***@***.***>
> Subject: Re: [openjournals/joss-reviews] [REVIEW]: MSG: A software package for interpolating stellar spectra in pre-calculated grids (Issue #4602)
>
>
> Thanks @ivastar<https://github.com/ivastar> for your response; I've committed a fix for the broken link, and it will appear when I tag a new patch release (1.1.1)
>
