### 2023-01-22 #4602 @rhdtownsend
Hi Arfon —

Done — the DOI is 10.5281/zenodo.7559319, and the full link is https://zenodo.org/record/7559319#.Y822CxPMLt0.

cheers,

Rich

> On Jan 22, 2023, at 2:06 AM, Arfon Smith ***@***.***> wrote:
>
>
> @rhdtownsend – At this point could you make a new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
>     • The title of the archive is the same as the JOSS paper title
>     • That the authors of the archive are the same as the JOSS paper authors
> I can then move forward with accepting the submission.
